Available lab names

Names of the labs available in the package.

LAB_DETAILS

Format

LAB_DETAILS

A data frame with 95 rows and 4 columns:

short_name

Short lab name

long_name

Long lab name

units

a list column with all the units available for the lab

default_units

the default units for the lab

low_male,high_male,low_female,high_female

the reference ranges for the lab, taken from the American Board of Internal Medicine. Can be NA if the lab does not have reference ranges.

Source

American Board of Internal Medicine. ABIM Laboratory Test Reference Ranges — July 2021. https://www.abim.org/~/media/ABIM%20Public/Files/pdf/exam/laboratory-reference-ranges.pdf (2021).

Examples

head(LAB_DETAILS)
#>   short_name                     long_name
#> 1        WBC       White Blood Cells (WBC)
#> 2        RBC         Red Blood Cells (RBC)
#> 3 Hemoglobin                    Hemoglobin
#> 4 Hematocrit                    Hematocrit
#> 5  Platelets                     Platelets
#> 6        MCV Mean Corpuscular Volume (MCV)
#>                                   units default_units low_male high_male
#> 1 x10E3/uL, x10E6/uL, x10E9/L, x10E12/L      x10E3/uL      4.4      11.0
#> 2 x10E6/uL, x10E3/uL, x10E9/L, x10E12/L      x10E6/uL      4.5       5.9
#> 3              mmol/L, g/dL, g/L, mg/mL          g/dL     14.0      18.0
#> 4                   ratio units, L/L, %             %     42.0      50.0
#> 5 x10E3/uL, x10E6/uL, x10E9/L, x10E12/L      x10E3/uL    150.0     450.0
#> 6                                    fL            fL     80.0      96.0
#>   low_female high_female
#> 1        4.4        11.0
#> 2        4.1         5.1
#> 3       12.0        16.0
#> 4       37.0        47.0
#> 5      150.0       450.0
#> 6       80.0        96.0