Download high-resolution reference distributions

The data is downloaded to the directory specified by the dir parameter. Note that if you specified a directory different from the default, you will need to set options(labNorm.dir = dir) in order for the package to use the downloaded data in future sessions.
Default directories are:

Unix: ~/.local/share/LabNorm

Mac OS X: ~/Library/Application Support/LabNorm

Win XP (not roaming): C:\\Documents and Settings\\<username>\\Data\\<AppAuthor>\\LabNorm

Win XP (roaming): C:\\Documents and Settings\\<username>\\Local Settings\\Data\\<AppAuthor>\\LabNorm

Win 7 (not roaming): C:\\Users\\<username>\\AppData\\Local\\<AppAuthor>\\LabNorm

Win 7 (roaming): C:\\Users\\<username>\\AppData\\Roaming\\<AppAuthor>\\LabNorm

ln_download_data(dir = NULL)

ln_data_downloaded()

Arguments

dir

the directory to download the data to. If NULL and the user approves, the data will be downloaded to the package directory, using rappdirs::user_data_dir("labnorm"), otherwise - a temporary directory would be used.

Value

None.

True if the data was downloaded, false otherwise.

Examples

# \donttest{
ln_download_data()
#> → Downloading to a temporary directory /tmp/RtmpyATdZH.
#> → Extracting data to /tmp/RtmpyATdZH.
#> → Extracting data to /tmp/RtmpyATdZH.
#> ✔ Data downloaded successfully.
# }

ln_data_downloaded()
#> [1] TRUE