Normalize lab values to age and gender

Normalize standard laboratory measurements (e.g. hemoglobin, cholesterol levels) according to age and gender, based on the algorithms described in "Personalized lab test models to quantify disease potentials in healthy individuals" doi:10.1038/s41591-021-01468-6.
The reference distribution used in this function are based on 2.1B lab measurements taken from 2.8M individuals between 2002-2019, filtered to exclude severe chronic diseases and medication effects. The resulting normalized value is a quantile between 0 and 1, representing the value's position in the reference distribution.
The list of supported labs can be found below or by running LAB_INFO$short_name.

ln_normalize(values, age, sex, lab, units = NULL)

ln_normalize_multi(df)

Arguments

values

a vector of lab values

age

a vector of ages between (20-89). Can be a single value if all values are the same age.

sex

a vector of either "male" or "female". Can be a single value if all values are the same sex.

lab

the lab name. See LAB_INFO$short_name for a list of available labs.

units

the units of the lab values. See ln_lab_units(lab) for a list of available units for each lab. If NULL then the default units (ln_lab_default_units(lab)) for the lab will be used. If different values have different units then this should be a vector of the same length as values.

df

a data frame with the columns "value", "age", "sex", "units", and "lab". The "lab" column should be a vector with the lab name per row. See ln_normalize for details on the other columns.

Value

a vector of normalized values. If ln_download_data() was not run, a lower resolution reference distribution will be used, which can have an error of up to 5 quantiles (0.05). Otherwise, the full reference distribution will be used. You can check if the high resolution data was downloaded using ln_is_high_res().

If the quantile information is not available (e.g. "Estradiol" for male patients), then the function will return NA.

reference distribution

The reference distribution used in the function has a resolution of 20 quantile bins and therefore may have an error of up to 5 quantiles (0.05), particularly at the edges of the distribution. The full reference distributions can be used after downloading the data using the ln_download_data() function.
The function would first look for the downloaded values at getOption("labNorm.dir"), then at rappdirs::user_data_dir("Labnorm"), then at the current working directory. If the data is not found, the lower resolution distribution.
You can check if the high resolution data was downloaded using ln_is_high_res().

labs

The following labs are supported:

WBC

White Blood Cells (WBC)

RBC

Red Blood Cells (RBC)

Hemoglobin

Hematocrit

Platelets

MCV

Mean Corpuscular Volume (MCV)

MCH

Mean Corpuscular Hemoglobin (MCH)

MCHC

Mean Corpuscular Hemoglobin Concentration (MCHC)

RDW

Red cell Distribution Width (RDW)

MPV

Mean Platelet Volume (MPV)

Albumin

Total Cholesterol

Triglycerides

BMI

Body Mass Index (BMI)

Iron

Transferrin

Ferritin

Total Globulin

Fibrinogen

Lymphocytes, Abs

Lymphocytes, Absolute

Lymphocytes, %

Neutrophils, Abs

Neutrophils, Absolute

Neutrophils, %

Monocytes, Abs

Monocytes, Absolute

Monocytes, %

Eosinophils, Abs

Eosinophils, Absolute

Eosinophils, %

Basophils, Abs

Basophils, Absolute

Basophils, %

Glucose

Glucose, Fasting

Urea

Creatinine

Uric Acid

Calcium

Phosphorus

Total Protein

HDL Cholesterol

LDL Cholesterol

Alk. Phosphatase

Alkaline Phosphatase (AP)

AST

Aspartate aminotransferase (AST)

ALT

Alanine aminotransferase (ALT)

GGT

Gamma-glutamyltransferase (GGT)

LDH

Lactate dehydrogenase (LDH)

CPK

Creatine Phospho-Kinase (CPK)

Total Bilirubin

Direct Bilirubin

Hemoglobin A1c

Sodium

Potassium

Vitamin D (25-OH)

Vitamin D, 25-Hydroxy

TSH

Thyroid-Stimulating Hormone (TSH)

T3, Free

Triiodothyronine (T3), Free

T4, Free

Thyroxine (T4), Free (Direct)

Blood Pressure, Systolic

Systolic Blood Pressure

Blood Pressure, Diastolic

Diastolic Blood Pressure

Vitamin B12

PSA

Prostate Specific Antigen (PSA)

ESR

Erythrocyte Sedimentation Rate (ESR)

CRP

C-Reactive Protein, Quant (CRP)

Amylase

Folic Acid

Magnesium

Indirect Bilirubin

LH

Luteinizing Hormone (LH)

Estradiol

Estradiol (E2)

Examples

# Normalize Hemoglobin values to age and sex
hemoglobin_data$quantile <- ln_normalize(
    hemoglobin_data$value,
    hemoglobin_data$age,
    hemoglobin_data$sex,
    "Hemoglobin"
)
#> The high resolution reference distributions are not available. Would you like to download them now?
#> (this question will only be asked once per session)
#> 1: Yes
#> 2: No
#> 
#> Error in the$yesno2("The high resolution reference distributions are not available. Would you like to download them now?\n(this question will only be asked once per session)"): user choose to exit

# plot the quantiles vs values for age 50-60
library(ggplot2)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
hemoglobin_data %>%
    filter(age >= 50 & age <= 60) %>%
    ggplot(aes(x = value, y = quantile, color = sex)) +
    geom_point() +
    theme_classic()
#> Don't know how to automatically pick scale for object of type <function>. Defaulting to
#> continuous.
#> Error in geom_point(): Problem while computing aesthetics.
#> ℹ Error occurred in the 1st layer.
#> Caused by error in `compute_aesthetics()`:
#> ! Aesthetics are not valid data columns.
#> ✖ The following aesthetics are invalid:
#> ✖ `y = quantile`
#> ℹ Did you mistype the name of a data column or forget to add `after_stat()`?

# Different units
hemoglobin_diff_units <- hemoglobin_data
hemoglobin_diff_units$value <- hemoglobin_diff_units$value * 0.1
hemoglobin_diff_units$quantile <- ln_normalize(
    hemoglobin_data$value,
    hemoglobin_data$age,
    hemoglobin_data$sex,
    "Hemoglobin",
    "mg/mL"
)
#> ℹ Converting mg/mL to g/dL for lab Hemoglobin. Using the formula `0.1 * x`.
#> → Using default quantiles. For higher resolution quantiles, run `ln_download_data()`. This message will only be shown once per session.

# Multiple units
creatinine_diff_units <- creatinine_data
creatinine_diff_units$value <- c(
    creatinine_diff_units$value[1:500] * 0.011312,
    creatinine_diff_units$value[501:1000] * 11.312
)
creatinine_diff_units$quantile <- ln_normalize(
    creatinine_diff_units$value,
    creatinine_diff_units$age,
    creatinine_diff_units$sex,
    "Creatinine",
    c(rep("umol/L", 500), rep("mmol/L", 500))
)
#> ℹ Converting umol/L to mg/dL for lab Creatinine. Using the formula `0.011312 * x`.
#> ℹ Converting mmol/L to mg/dL for lab Creatinine. Using the formula `11.312 * x`.

library(dplyr)
multi_labs_df <- bind_rows(
    hemoglobin_data %>% mutate(lab = "Hemoglobin"),
    creatinine_data %>% mutate(lab = "Creatinine")
)

multi_labs_df$quantile <- ln_normalize_multi(multi_labs_df)

head(multi_labs_df)
#>   age    sex value        lab   quantile
#> 1  20   male  9.39 Hemoglobin 0.03838718
#> 2  20   male 14.03 Hemoglobin 0.18914947
#> 3  20   male 14.44 Hemoglobin 0.28589076
#> 4  20   male 15.80 Hemoglobin 0.75032395
#> 5  20 female 12.06 Hemoglobin 0.24222206
#> 6  20 female 12.89 Hemoglobin 0.55270091