Compute the lab value for a given quantile

This function computes the lab value for given quantiles in the default units for the lab. Default units can be found using the function ln_lab_default_units. In case where no quantile is available for a given lab, age, and sex the function returns NA.
Note that the values of very high or low quantiles (e.g. >0.95,<0.05 on the low resolution version, >0.99,<0.01 on the high resolution version) are not reliable as they can represent technical outliers of the data.

ln_quantile_value(
  quantiles,
  age,
  sex,
  lab,
  reference = "Clalit",
  allow_edge_quantiles = FALSE
)

Arguments

quantiles

a vector of quantiles (in the range 0-1) to compute the lab value for.

age

a vector of ages between 20-89 for "Clalit" reference and 35-80 for "UKBB". Can be a single value if all values are the same age.

sex

a vector of either "male" or "female". Can be a single value if all values are the same sex.

lab

The lab name.

reference

the reference distribution to use. Can be either "Clalit" or "UKBB" or "Clalit-demo". Please download the Clalit and UKBB reference distributions using ln_download_data().

allow_edge_quantiles

If TRUE (default) then the function will return the value for the edge quantiles (<0.01 or >0.99) even though they are not reliable. If FALSE then the function will return NA for those quantiles. Note that for the "Clalit-demo" reference, the threshold would be <0.05 or >0.95.

Value

a data frame which contains the values for each combination of quantile, age and sex. The data frame has the the following columns:

age: age in years

sex: "male" or "female"

quantile: he quantile

value: the lab value

unit: the lab unit

lab: the lab name

Examples

# \donttest{
ln_quantile_value(c(0.05, 0.5, 0.95), 50, "male", "WBC")
#>   age  sex quantile     value     unit lab
#> 1  50 male     0.05  4.584943 x10E3/uL WBC
#> 2  50 male     0.50  7.182971 x10E3/uL WBC
#> 3  50 male     0.95 11.789079 x10E3/uL WBC

ln_quantile_value(
    c(0, 0.05, 0.1, 0.4, 0.5, 0.6, 0.9, 1),
    c(50, 60),
    c("male", "female"),
    "Glucose"
)
#>    age    sex quantile     value  unit     lab
#> 1   50   male     0.00        NA mg/dL Glucose
#> 2   50   male     0.05  76.63210 mg/dL Glucose
#> 3   50   male     0.10  80.92743 mg/dL Glucose
#> 4   50   male     0.40  90.84690 mg/dL Glucose
#> 5   50   male     0.50  92.98665 mg/dL Glucose
#> 6   50   male     0.60  95.83511 mg/dL Glucose
#> 7   50   male     0.90 111.93415 mg/dL Glucose
#> 8   50   male     1.00        NA mg/dL Glucose
#> 9   60   male     0.00        NA mg/dL Glucose
#> 10  60   male     0.05  77.98954 mg/dL Glucose
#> 11  60   male     0.10  81.92832 mg/dL Glucose
#> 12  60   male     0.40  92.60682 mg/dL Glucose
#> 13  60   male     0.50  94.97742 mg/dL Glucose
#> 14  60   male     0.60  97.86284 mg/dL Glucose
#> 15  60   male     0.90 113.95915 mg/dL Glucose
#> 16  60   male     1.00        NA mg/dL Glucose
#> 17  50 female     0.00        NA mg/dL Glucose
#> 18  50 female     0.05  75.29256 mg/dL Glucose
#> 19  50 female     0.10  78.77304 mg/dL Glucose
#> 20  50 female     0.40  87.97072 mg/dL Glucose
#> 21  50 female     0.50  90.91937 mg/dL Glucose
#> 22  50 female     0.60  92.91284 mg/dL Glucose
#> 23  50 female     0.90 106.98740 mg/dL Glucose
#> 24  50 female     1.00        NA mg/dL Glucose
#> 25  60 female     0.00        NA mg/dL Glucose
#> 26  60 female     0.05  77.70802 mg/dL Glucose
#> 27  60 female     0.10  80.54148 mg/dL Glucose
#> 28  60 female     0.40  90.41370 mg/dL Glucose
#> 29  60 female     0.50  92.59323 mg/dL Glucose
#> 30  60 female     0.60  95.86970 mg/dL Glucose
#> 31  60 female     0.90 110.90176 mg/dL Glucose
#> 32  60 female     1.00        NA mg/dL Glucose
# }

# \dontshow{
ln_quantile_value(c(0.05, 0.5, 0.95), 50, "male", "WBC", reference = "Clalit-demo")
#>   age  sex quantile     value     unit lab
#> 1  50 male     0.05  4.584943 x10E3/uL WBC
#> 2  50 male     0.50  7.182971 x10E3/uL WBC
#> 3  50 male     0.95 11.789079 x10E3/uL WBC
# }